Suggestion to improve tweet - split the values that are accumulated and always growing - confirmados, recuperados, obitos - from the values that are daily snapshots - ativos, internados, uci - to prevent further confusion of "why is confirmados_novos the sum of…" - remove values for 'recuperados', keeping only the percentage at the end - add breath space in between value(other) - third tweet just saying "all data on github" without mentioning what or when. There are a lot more data there - population and sample code
This way we have a clear "novos casos" and "novos obitos", two values always growing, with the percentage of growth, and the total. Not recuperados here.

Then, separated by an empty line, the three (out of four) daily snapshot values: activos, (from which some are) internados, (from which some are) UCI. Enfermaria is also skipped.
We want these three to go to zero.

Then it keeps the thumbs up for the recuperados as a percentage.
We want this to go to 100%